# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Memora API (mock)
  version: "1.0.0"
  description: |
    API mock do Memora — Express 5, estado em memória (reinicia com o processo, sem banco de
    dados). Reproduz as regras de negócio do Escopo Consolidado v2 para servir de backend ao
    protótipo navegável; não é a API de produção.

    Todas as rotas de negócio vivem sob `/v1`. A autenticação é por sessão simples (token opaco
    em `Authorization: Bearer <token>`), sem JWT nem refresh — ver
    [ADR-0008](./adr/0008-autenticacao-por-sessao-simples.md).

    Cada área do produto (aluno, responsável, B2B, back-office, parceiro) faz login com sua
    própria rota `POST .../auth/login` e recebe um token cujo `role` restringe o que pode
    acessar (`requireRole` em `backend/src/lib/auth.js`).
  contact:
    name: Memora — engenharia
  license:
    name: Uso interno
servers:
  - url: http://localhost:3180/v1
    description: Desenvolvimento local (porta configurável via `PORT`)
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: >
        Token de sessão retornado por qualquer rota `auth/login`. Sessões expiram em 30 dias
        sem uso (deslizante); usar em `me`, `parental`, `b2b`, `admin` e `partner` é mutuamente
        exclusivo por `role`.
  parameters:
    deckId:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: ID do deck.
...
  responses:
...
    NotFound:
      description: Recurso não encontrado (ou não pertence ao usuário autenticado).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Validation:
      description: Corpo da requisição inválido.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, description: "Slug estável da causa, ex.: validation, not_found, guardian_required." }
        message: { type: string, description: "Mensagem em PT-BR, pronta para exibição." }
      additionalProperties: true
      description: Formato uniforme de erro em toda a API (ver `fail()` em `backend/src/lib/helpers.js`).
...
    Comment:
      type: object
      properties:
        id: { type: string }
        deckId: { type: string }
        authorId: { type: string }
        authorName: { type: string }
        text: { type: string }
        createdAt: { type: string, format: date-time }
        replies:
          type: array
          items: { $ref: '#/components/schemas/Comment' }
...
    Report:
      type: object
      properties:
        id: { type: string }
        targetType: { type: string, enum: [deck, comment, contato_suspeito] }
        targetId: { type: string }
        targetName: { type: string }
        reason: { type: string }
        detail: { type: string, nullable: true }
        reporter: { type: string }
        status: { type: string, enum: [aberta, auto_analise, resolvida] }
        createdAt: { type: string, format: date-time }
        resolution: { type: object, nullable: true, additionalProperties: true }
...
paths:
...
  /community/search:
    get:
      tags: [Comunidade]
      summary: Busca de decks públicos
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: category, in: query, schema: { type: string } }
        - { name: difficulty, in: query, schema: { type: string, enum: [facil, medio, dificil] } }
        - { name: sort, in: query, schema: { type: string, enum: [relevance, recent, rating] } }
        - { name: approved, in: query, schema: { type: string, enum: ["1"] }, description: "Filtra por criadores com selo." }
      responses: { "200": { description: OK } }
  /community/decks/{id}:
    get:
      tags: [Comunidade]
      summary: Detalhe do deck público, cards, tradução e versões
      parameters: [{ $ref: '#/components/parameters/deckId' }]
      responses:
        "200": { description: OK }
        "404": { $ref: '#/components/responses/NotFound' }
  /community/decks/{id}/favorite:
    post:
      tags: [Comunidade]
      summary: Alterna favorito (com janela de desfazer de 3 s no cliente)
      parameters: [{ $ref: '#/components/parameters/deckId' }]
      responses: { "200": { description: OK } }
  /community/decks/{id}/clone:
    post:
      tags: [Comunidade]
      summary: Clona o deck para "Meus decks" (paga 5 Neurônios de bônus ao autor)
      parameters: [{ $ref: '#/components/parameters/deckId' }]
      responses: { "201": { description: OK } }
  /community/decks/{id}/vote:
    post:
      tags: [Comunidade]
      summary: Vota no deck (exige ≥ 30% dos cards ou 20 cards respondidos)
      parameters: [{ $ref: '#/components/parameters/deckId' }]
      responses:
        "200": { description: OK }
        "403": { description: vote_not_allowed }
  /community/decks/{id}/react:
    post:
      tags: [Comunidade]
      summary: "Reage com um emoji fixo (👏🔥🧠💡❤️😂)"
      parameters: [{ $ref: '#/components/parameters/deckId' }]
      requestBody: { content: { application/json: { schema: { type: object, required: [emoji], properties: { emoji: { type: string } } } } } }
      responses:
        "200": { description: OK }
        "422": { $ref: '#/components/responses/Validation' }
  /community/decks/{id}/share:
    post:
      tags: [Comunidade]
      summary: Gera link público de compartilhamento (exige liberação do responsável se menor)
      parameters: [{ $ref: '#/components/parameters/deckId' }]
      responses:
        "200": { description: OK }
        "403": { description: guardian_required }
  /community/decks/{id}/report:
    post:
      tags: [Comunidade]
      summary: Denuncia o deck
      parameters: [{ $ref: '#/components/parameters/deckId' }]
      requestBody: { content: { application/json: { schema: { type: object, required: [reason], properties: { reason: { type: string }, detail: { type: string } } } } } }
      responses:
        "201": { description: Registrada, content: { application/json: { schema: { type: object, properties: { report: { $ref: '#/components/schemas/Report' } } } } } }
        "422": { $ref: '#/components/responses/Validation' }
  /community/decks/{id}/translate:
    post:
      tags: [Comunidade]
      summary: Traduz os cards do deck (cache por conteúdo+idioma; original nunca muda)
      parameters: [{ $ref: '#/components/parameters/deckId' }]
      responses: { "200": { description: OK } }
  /community/decks/{id}/comments:
    get:
      tags: [Comunidade]
      summary: Lista comentários (com respostas aninhadas)
      parameters: [{ $ref: '#/components/parameters/deckId' }]
      responses: { "200": { description: OK, content: { application/json: { schema: { type: object, properties: { threads: { type: array, items: { $ref: '#/components/schemas/Comment' } } } } } } } }
    post:
      tags: [Comunidade]
      summary: Comenta ou responde (exige verificação e liberação do responsável se menor)
      parameters: [{ $ref: '#/components/parameters/deckId' }]
      requestBody: { content: { application/json: { schema: { type: object, required: [text], properties: { text: { type: string }, parentId: { type: string, nullable: true } } } } } }
      responses:
        "201": { description: Criado }
        "403": { description: "verification_required · guardian_required" }
        "422": { $ref: '#/components/responses/Validation' }
  /community/comments/{id}/report:
    post:
      tags: [Comunidade]
      summary: Denuncia um comentário (vai para análise automática)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "201": { description: OK } }
  /community/authors/{id}:
    get:
      tags: [Comunidade]
      summary: Perfil público de um criador e seus decks
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: OK }
        "404": { $ref: '#/components/responses/NotFound' }
  /community/authors/{id}/follow:
    post:
      tags: [Comunidade]
      summary: Segue/deixa de seguir o criador (exige liberação do responsável se menor)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: OK }
        "403": { description: guardian_required }
        "404": { $ref: '#/components/responses/NotFound' }
...
  /community/decks/{id}/study-start:
    post:
      tags: [Comunidade]
      summary: Inicia estudo de um deck da comunidade sem clonar (progresso registrado por resposta)
      parameters: [{ $ref: '#/components/parameters/deckId' }]
      responses: { "200": { description: OK } }

  # ---------------------------------------------------------------- Carteira
...